added the vpn agent and driver test cases
